base_agent/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tether-agent"
version = "1.1.0"
dependencies = ["paho_mqtt==1.6.1", "msgpack==1.0.8"]
requires-python = ">= 3.11"
